Refactor of the ProductServiceTest.
