+++ OnlineShop.UnitTests/Helpers/TestAsyncEnumerator.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OnlineShop.UnitTests.Helpers
+{
+    public class TestAsyncEnumerator<T> : IAsyncEnumerator<T>
+    {
+        private readonly IEnumerator<T> _enumerator;
+
+        public TestAsyncEnumerator(IEnumerator<T> enumerator)
+        {
+            _enumerator = enumerator;
+        }
+
+        public T Current => _enumerator.Current;
+
+        public ValueTask DisposeAsync()
+        {
+            _enumerator.Dispose();
+            return new ValueTask(Task.CompletedTask);
+        }
+
+        public ValueTask<bool> MoveNextAsync()
+        {
+            return new ValueTask<bool>(_enumerator.MoveNext());
+        }
+    }
+}
